% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_pathwaysPCA.R
\name{avg.pathwaysPCA}
\alias{avg.pathwaysPCA}
\title{Average pathways of post-disturbance dynamics (PCA)}
\usage{
avg.pathwaysPCA(
  abundance,
  sites,
  surveys,
  cluster,
  type = "mixed",
  onlyPaths = NULL,
  lwd.factor = NULL,
  axes = c(1, 2),
  species = FALSE,
  sp.fac = 1,
  zoom = FALSE,
  add.legend = FALSE,
  cluster.name = NULL,
  ...
)
}
\arguments{
\item{abundance}{A data frame containing species abundance of target communities
for three states relative to the occurrence of a disturbance (pre-disturbance,
disturbed, and post-disturbance states). The three surveys of each site are in
rows and species are in columns.}

\item{sites}{A vector indicating the site of each community state.}

\item{surveys}{A vector indicating the survey of each community state.}

\item{cluster}{A numeric or a vector of colors indicating the cluster to which
each community state belongs or the color used for each state.}

\item{type}{A character value indicating the type of pathways to be plotted.
\code{"mixed"} circular dots representing pre-disturbance states are linked by
lines to post-disturbance states represented by triangles. \code{"points"}
pre-disturbance, disturbed and post-disturbance states are represented by circular,
squared and triangular points, respectively. \code{"arrows"} post-disturbance
dynamics are represented by arrows.}

\item{onlyPaths}{A vector of length three containing the cluster identifier of
each state (i.e. pre-disturbance, disturbed, and post-disturbance states) to be
plotted. If all clusters are considered for one or two states, corresponding value/s
must be replaced by \code{NA}. If all pathways must be considered, set \code{onlyPaths = NULL}.}

\item{lwd.factor}{A positive number to use line widths proportional to the number
of plots following the same pathways. The default \code{NULL} uses the same \code{lwd = 1}
for all pathways.}

\item{axes}{The pair of the principal coordinates to be used in the plot.}

\item{species}{A logical value indicating if species arrows will be plotted.}

\item{sp.fac}{Factor by which species arrows will be multiplied (only for graphical purposes).}

\item{zoom}{A logical value to zoom the plot to the community state coordinates.}

\item{add.legend}{A logical flag to add the plot legend.}

\item{cluster.name}{A character vector indicating the name of the cluster to which each
community state belongs. Values are specified in the legend if \code{add.legend = TRUE}.}

\item{...}{Additional parameters to \code{plot}}
}
\value{
Average post-disturbance dynamics of a set of communities are displayed
as pathways represented in an ordination space (PCA) by linking the pre-disturbance,
disturbed, and post-disturbance states in chronological order.
}
\description{
The function \code{avg.pathwaysPCA} summarizes and represents average
post-disturbance dynamics of a set of disturbed communities as pathways displayed
in an ordination space resulting from applying principal component analyses (PCA)
on an abundance matrix.
}
\details{
Average pathways are constructed by linking in chronological order the
centroids calculated for each group of pre-disturbance, disturbed, and post-disturbance
states of the set of communities following equivalent dynamics (i.e. exhibiting
the same combination of clusters along the three states). Clusters corresponding
to each state and community must be previously defined and provided through the
argument \code{clusters}. For \code{type = "mixed} or \code{type = "points}
pre-disturbance, disturbed, and post-disturbance average states are represented
by the color corresponding to \code{cluster}.
}
\examples{

data("post_disturbance")

# Display individual pathways using a different color for each
# cluster
head(post_disturbance$cluster)
pathwaysPCA(abundance = post_disturbance$abundance,
         sites = post_disturbance$sites,
         surveys = post_disturbance$surveys,
         cluster = post_disturbance$cluster,
         type = "mixed",
         main = "Individual pathways",
         zoom = TRUE)

# Summarize individual into average pathways
avg.pathwaysPCA(abundance = post_disturbance$abundance,
             sites = post_disturbance$sites,
             surveys = post_disturbance$surveys,
             cluster = post_disturbance$cluster,
             type = "mixed",
             main = "Average pathways",
             lwd.factor = 20,
             zoom = TRUE)

# Display average dynamics of communities classified in cluster 4
# in the pre-disturbance state.
avg.pathwaysPCA(abundance = post_disturbance$abundance,
             sites = post_disturbance$sites,
             surveys = post_disturbance$surveys,
             cluster = post_disturbance$cluster,
             type = "mixed",
             onlyPaths = c(4, NA, NA),
             main = "Average pathways of 'cluster 4'",
             lwd.factor = 5,
             zoom = TRUE)


}
\references{
Sánchez-Pinillos, M., Leduc, A., Ameztegui, A. Kneeshaw, D., Lloret, F.,
Coll, L. 2019. Resistance, resilience or change: post-disturbance dynamics of boreal forests
after insect outbreaks. Ecosystems. https://doi.org/10.1007/s10021-019-00378-6
}
\seealso{
\code{\link{pathwaysPCA}}
}
\author{
Martina Sánchez-Pinillos (Forest Sciences Centre of Catalonia)
\email{martina.sanchezpinillos@gmail.com}
}
