% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwaysPCA.R
\name{pathwaysPCA}
\alias{pathwaysPCA}
\title{Individual pathways of post-disturbance dynamics (PCA)}
\usage{
pathwaysPCA(
  abundance,
  sites,
  surveys,
  type = "mixed",
  onlyDraw = NULL,
  cluster = NULL,
  axes = c(1, 2),
  species = FALSE,
  sp.fac = 1,
  zoom = FALSE,
  add.legend = FALSE,
  cluster.name = NULL,
  ...
)
}
\arguments{
\item{abundance}{A data frame containing species abundance of target communities for
three states relative to the occurrence of a disturbance (pre-disturbance, disturbed,
and final states). The three surveys of each site are in rows and species
are in columns.}

\item{sites}{A vector indicating the site of each community state.}

\item{surveys}{A vector indicating the survey of each community state.}

\item{type}{A character value indicating the type of pathways to be plotted.
\code{"mixed"} circular dots representing pre-disturbance states are linked by lines
to final states represented by triangles. \code{"points"} pre-disturbance,
disturbed and post-disturbance states are represented by circular, squared and triangular
points, respectively. \code{"arrows"} post-disturbance dynamics are represented by arrows.}

\item{onlyDraw}{A vector indicating a subset of sites to be plotted. Sites' identifier
must match to those in \code{sites}.}

\item{cluster}{A numeric or a vector of colors indicating the cluster to which each
community state belongs or the color used for each state. This option is only available
for \code{type = "mixed"} and \code{type = "points"}}

\item{axes}{The pair of the principal coordinates to be used in the plot.}

\item{species}{A logical flag to draw species arrows.}

\item{sp.fac}{Factor by which species arrows will be multiplied (only for graphical purposes).}

\item{zoom}{A logical flag to zoom the plot to the community state coordinates.}

\item{add.legend}{A logical flag to add the plot legend.}

\item{cluster.name}{A character vector indicating the name of the cluster to which each
community state belongs. Values are specified in the legend if \code{add.legend = TRUE}.}

\item{...}{Additional parameters to \code{plot}}
}
\value{
Individual post-disturbance dynamics of a given set of communities are displayed
as pathways in an ordination space (PCA) by linking the pre-disturbance, disturbed, and
final states in chronological order.
}
\description{
The function \code{pathwaysPCA} represents post-disturbance dynamics of a set
of disturbed communities as pathways displayed in an ordination space resulting from
applying principal component analyses (PCA) on the abundance matrix.
}
\examples{

data("post_disturbance")

# Display pre-disturbance, disturbed, and final states as orange,
# blue, and red points, respectively, for a set of selected sites.
color = rep(c("orange", "blue", "red"),
            length(unique(post_disturbance$sites)))
selected.sites = c(3, 17, 66, 86, 132, 157)
pathwaysPCA(abundance = post_disturbance$abundance,
         sites = post_disturbance$sites,
         surveys = post_disturbance$surveys,
         cluster = color,
         type = "points", main = 'type = "points"',
         onlyDraw = selected.sites,
         zoom = TRUE)

# type = "mixed"
pathwaysPCA(abundance = post_disturbance$abundance,
         sites = post_disturbance$sites,
         surveys = post_disturbance$surveys,
         cluster = color,
         type = "mixed", main = 'type = "mixed"',
         onlyDraw = selected.sites,
         zoom = TRUE)

# If type = "arrows" a different color is used to represent
# post-disturbance dynamics of each site
pathwaysPCA(abundance = post_disturbance$abundance,
         sites = post_disturbance$sites,
         surveys = post_disturbance$surveys,
         type = "arrows", main = 'type = "arrows"',
         onlyDraw = selected.sites,
         zoom = TRUE)




}
\references{
Sánchez-Pinillos, M., Leduc, A., Ameztegui, A. Kneeshaw, D., Lloret, F.,
Coll, L. 2019. Resistance, resilience or change: post-disturbance dynamics of boreal forests
after insect outbreaks. Ecosystems. https://doi.org/10.1007/s10021-019-00378-6
}
\seealso{
\code{\link{avg.pathwaysPCA}}
}
\author{
Martina Sanchez-Pinillos (Forest Sciences Centre of Catalonia)
\email{martina.sanchezpinillos@gmail.com}
}
