% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{resistance}
\alias{resistance}
\title{Resistance index}
\usage{
resistance(d, sites, surveys)
}
\arguments{
\item{d}{a symmetric \code{matrix} or an object of class \code{dist} cointaining
the dissimilarity values between pairs of community states. \code{d} values
must be in the range [0, 1].}

\item{sites}{A vector indicating the site of each community state.}

\item{surveys}{A vector indicating the survey of each community state.}
}
\value{
A data frame cointaining the resistance value (Rt) of each site.
}
\description{
The function \code{resistance} measures how stable a community is
during a given disturbance in terms of a pre-defined dissimilarity metric.
}
\details{
The resistance index is calculated as the one-complement of a
dissimilarity measure calculated to compare two community states: the
pre-disturbance state (i.e. before disturbance)and the disturbed state
(i.e. during or immediately after disturbance). The resistance index
ranges between 0 (maximum dissimilarity) and 1 (identical pre-disturbance
and disturbed states).
For that, the dissimilarity measure used to generate \code{d} must be
bounded between 0 and 1. Otherwise, it must be previously transformed.
If more than two surveys are provided for a given site, only the
first and second surveys are considered as the pre-disturbance and disturbed
states, respectively, to calculate resistance.
}
\references{
Sánchez-Pinillos, M., Leduc, A., Ameztegui, A. Kneeshaw, D., Lloret, F.,
Coll, L. Resistance, resilience or change: post-disturbance dynamics of boreal forests
after insect outbreaks. (Submitted to Ecosystems)
}
\seealso{
\code{\link{resilience}}
}
\author{
Martina Sánchez-Pinillos (Forest Sciences Centre of Catalonia)
\email{martina.sanchezpinillos@gmail.com}
}
